Create event and event listener for update cache in user-based recommendation

+++ src/main/java/com/fleur/cinemate/listener/RecordDeleteEventListener.java
@@ -4,30 +4,32 @@ import com.fleur.cinemate.event.RecordDeletedEvent;
 import com.fleur.cinemate.pending.PendingDelete;
 import com.fleur.cinemate.pending.PendingDeleteRepository;
 import com.fleur.cinemate.search.service.ESDeleteService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+@RequiredArgsConstructor
 @Log4j2
 @Component
 public class RecordDeleteEventListener {
     private final ESDeleteService esDeleteService;
     private final PendingDeleteRepository pendingDeleteRepository;
 
-    public RecordDeleteEventListener(ESDeleteService esDeleteService, PendingDeleteRepository pendingDeleteRepository) {
-        this.esDeleteService = esDeleteService;
-        this.pendingDeleteRepository = pendingDeleteRepository;
-    }
-
     @Async
     @EventListener
     @Transactional
-    public void handleRecordDeletedEvent(RecordDeletedEvent recordDeletedEvent) {
-        Long recordId = recordDeletedEvent.getRecordId();
-        String entityName = recordDeletedEvent.getEntityName();
-
+    public void handleRecordDeleted(RecordDeletedEvent event) {
+        if (event == null || event.getRecordId() == null
+                || event.getEntityName() == null
+                || event.getEntityName().isBlank()) {
+            log.warn("RecordDeleted event is null or empty");
+            return;
+        }
+        Long recordId = event.getRecordId();
+        String entityName = event.getEntityName();
         try {
             esDeleteService.deleteByEntityName(entityName, recordId);
             log.info("Record {} with id: {} has been deleted", entityName, recordId);
